backend/pyproject.toml
[project]
name = "interview-guide"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "bs4>=0.0.2",
    "chromadb>=1.0.20",
    "fastapi>=0.120.4",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.10",
    "langgraph>=0.6.6",
    "langsmith>=0.4.21",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.0",
    "sqlite-utils>=3.38",
    "streamlit>=1.49.1",
    "tavily-python>=0.7.11",
    "uvicorn[standard]>=0.35.0",
]
